Lightline : add first draft of sourcerer colorscheme
This is inspired by romainl/Apprentice (see the fancylines branch) to
get all the values of the colors.

I changed the format of the commands to use the lightline#flatten
function, allowing more readable commands in my opinion

Added the terminal section.

TODO : allow divergence between GUI and ANSI colors by setting the
original gui colors of sourcerer in the s:guiXX variables